Renamed types/db to types/dto
