+++ logicle/app/admin/assistants/[id]/page.tsx
@@ -7,7 +7,7 @@ import { useRouter } from 'next/navigation'
 import toast from 'react-hot-toast'
 import { useTranslation } from 'next-i18next'
 import { AssistantForm } from '../components/AssistantForm'
-import { InsertableAssistant, SelectableAssistantWithTools } from '@/types/db'
+import { InsertableAssistant, SelectableAssistantWithTools } from '@/types/dto'
 import { patch } from '@/lib/fetch'
 import { useSWRJson } from '@/hooks/swr'
 import { AdminPageTitle } from '../../components/AdminPageTitle'
