+++ logicle/app/admin/assistants/components/AssistantForm.tsx
@@ -17,7 +17,7 @@ import { zodResolver } from '@hookform/resolvers/zod'
 import { ChangeEvent, useEffect, useRef, useState } from 'react'
 import { OpenAIModel } from '@/types/openai'
 import { Textarea } from '@/components/ui/textarea'
-import { File, AssistantTool, InsertableAssistantWithTools, InsertableFile } from '@/types/db'
+import { File, AssistantTool, InsertableAssistantWithTools, InsertableFile } from '@/types/dto'
 import ImageUpload from '@/components/ui/ImageUpload'
 import { Switch } from '@/components/ui/switch'
 import { Upload, UploadList } from '@/components/app/upload'
