+++ logicle/app/admin/assistants/components/AssistantPreview.tsx
@@ -1,4 +1,4 @@
-import { InsertableAssistant } from '@/types/db'
+import { InsertableAssistant } from '@/types/dto'
 import { Chat } from '@/app/chat/components/Chat'
 import { Attachment, ConversationWithMessages } from '@/types/chat'
 import ChatPageContext, { ChatPageContextProps } from '@/app/chat/components/context'
