+++ logicle/app/admin/assistants/create/page.tsx
@@ -5,7 +5,7 @@ import { useRouter } from 'next/navigation'
 import toast from 'react-hot-toast'
 import { useTranslation } from 'next-i18next'
 import { AssistantForm } from '../components/AssistantForm'
-import { InsertableAssistant } from '@/types/db'
+import { InsertableAssistant } from '@/types/dto'
 import { post } from '@/lib/fetch'
 import { useBackends } from '@/hooks/backends'
 import { AdminPageTitle } from '../../components/AdminPageTitle'
