+++ logicle/app/admin/tools/page.tsx
@@ -7,7 +7,7 @@ import toast from 'react-hot-toast'
 import { WithLoadingAndError } from '@/components/ui'
 import { delete_ } from '@/lib/fetch'
 import { AdminPageTitle } from '@/app/admin/components/AdminPageTitle'
-import { ToolDTO } from '@/types/db'
+import { ToolDTO } from '@/types/dto'
 import DeleteButton from '../components/DeleteButton'
 import { Link } from '@/components/ui/link'
 import { useRouter } from 'next/navigation'
