+++ logicle/app/admin/workspaces/components/WorkspaceSettings.tsx
@@ -1,4 +1,4 @@
-import { Workspace } from '@/types/db'
+import { Workspace } from '@/types/dto'
 import { useTranslation } from 'next-i18next'
 import { useRouter } from 'next/navigation'
 import React from 'react'
