+++ logicle/app/api/user/prompts/[promptId]/route.ts
@@ -1,6 +1,6 @@
 import { getPrompt, deletePrompt, updatePrompt } from 'models/prompt' // Import the helper functions
 import ApiResponses from '@/api/utils/ApiResponses'
-import { Prompt } from '@/types/db'
+import { Prompt } from '@/types/dto'
 import { requireSession } from '@/app/api/utils/auth'
 
 export const dynamic = 'force-dynamic'
