+++ logicle/models/user.ts
@@ -1,6 +1,6 @@
 import { db } from 'db/database'
 import type { Session } from 'next-auth'
-import { InsertableUser, UpdateableUser } from '@/types/db'
+import { InsertableUser, UpdateableUser } from '@/types/dto'
 import { UserRoleId } from '@/types/user'
 import { hashPassword } from '@/lib/auth'
 import { nanoid } from 'nanoid'
